create "UserController" class it has the 2 functions: async auth fn and initDataVerify fn.
initDataVerify

it verifies on authenticity "initData" that comes from Telegram and has 2 checking logic.

1. if "initData" doesn't come (it is undefined) the fn throw the error
2. second checking logic verify "initData" via special "isValid" fn and if "initData"
isn't authentic it also throw the error

after all that verifications and no one throws error the initDataVerify fn return the parsed
"initData" via special "parse" fn to read the info that is in it.

async auth

if initDataVerify throws the error async auth cathes it and send 400 status code.
otherwise it follows this steps:
1. it creates or updates user via prisma client.
2. it creates jwt and keeps it in cockie that will be in response
3. it sends response that contains the authorized user and a message